refactor scripts to use srraf library
