--- Portals/_default/Skins/CLIENT_CODE/src/js/modules/on-resize.js
@@ -1,7 +1,0 @@
-import debounce from '../utils/debounce';
-
-function onResize(callback, wait = 250) {
-  window.addEventListener('resize', debounce(callback, wait));
-}
-
-export default onResize;
