+++ Portals/_default/Skins/CLIENT_CODE/src/js/utils/is-mobile-device.js
