+++ Portals/_default/Skins/CLIENT_CODE/src/js/utils/is-touch-device.js
